STbackend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
filelock==3.20.3
future==1.0.0
gevent==25.9.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
passlib==1.7.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-file==3.0.1
SQLAlchemy==2.0.45
starlette==0.50.0
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
zope.event==6.1
zope.interface==8.2
